refactor: move screenReaderInstructions and announcements to own files
